fixed my wall werror wextra errors

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: gstrauss <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/27 10:02:16 by gstrauss          #+#    #+#             */
-/*   Updated: 2019/05/29 12:48:51 by gstrauss         ###   ########.fr       */
+/*   Updated: 2019/05/30 08:18:38 by gstrauss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void *ft_memchr(const void *s, int c, size_t n)
 {
-	int i;
+	size_t i;
 	char *ss;
 	unsigned char cc;
 	cc = (unsigned char)c;
