+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: gstrauss <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/27 10:03:56 by gstrauss          #+#    #+#             */
-/*   Updated: 2019/05/29 12:49:29 by gstrauss         ###   ########.fr       */
+/*   Updated: 2019/05/30 08:20:29 by gstrauss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void *ft_memcpy(void  *dst, const void *src, size_t n)
 {
-	int i;
+	size_t i;
 	char *dest, *source;
 	
 	dest = (char *)dst;
