+++ ft_strnequ.c
@@ -6,7 +6,7 @@
 /*   By: gstrauss <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/29 12:44:51 by gstrauss          #+#    #+#             */
-/*   Updated: 2019/05/29 12:45:05 by gstrauss         ###   ########.fr       */
+/*   Updated: 2019/05/30 08:29:08 by gstrauss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_strnequ(char const *s1, char const *s2, size_t n)
 {
-	int i;
+	size_t i;
 	i = 0;
 	while(s1[i] == s2[i] && i < n)
 	{
