+++ ft_strnew.c
@@ -6,7 +6,7 @@
 /*   By: gstrauss <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/27 12:14:56 by gstrauss          #+#    #+#             */
-/*   Updated: 2019/05/29 12:45:46 by gstrauss         ###   ########.fr       */
+/*   Updated: 2019/05/30 08:29:24 by gstrauss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char *	ft_strnew(size_t size)
 {
-	int i;
+	size_t i;
 	i = 0;
 	char *new;
 	new = (char *) malloc(size * sizeof(char));
